+++ scripts/coredump/coredump_gdbserver.py
@@ -10,8 +10,8 @@ import os
 import socket
 import sys
 
-from parser.log_parser import CoredumpLogFile
-from parser.elf_parser import CoredumpElfFile
+from coredump_parser.log_parser import CoredumpLogFile
+from coredump_parser.elf_parser import CoredumpElfFile
 
 import gdbstubs
