+++ scripts/coredump/coredump_parser/__init__.py
