+++ scripts/coredump/coredump_parser/elf_parser.py
